Add files via upload
Problem Statement - Case Study 1
Case Study 1: Problem Statement
Imagine you are in 1977 and you work as an air crash investigation officer on the Spanish island of Tenerife. You receive some bad news that two Boeing 747 passenger jets, KLM Flight 4805 and Pan Am Flight 1736, collided on the runway in the airport, which has resulted in the death of 583 people.

You have been tasked with investigating this incident to find the possible root causes for the crash. 

Reading Material

We request you to go through the following link to understand the scenario of the airport and the disaster; it includes a detailed list of the events and the measures that were taken by the investigation board to ensure air safety for future air travel.

To understand the details of the accident, please go through the material listed below.

https://en.wikipedia.org/wiki/Tenerife_airport_disaster - Wikipedia article describing the accident in detail.
The aforementioned link provides details of the root causes of the accident.

Following are two additional links that can further aid your understanding of the common factors that contribute to air crashes.

http://aviationknowledge.wikidot.com/asi:tenerife-airport-disaster – Additional material.
https://goflightmedicine.com/tenerife-disaster/ - Additional material.
 
For the air crash investigation, the Human Factors Analysis and Classification System (HFACS) is a well-known model to analyse human factors in accidents. This will also provide you with an issue tree framework to analyse air crashes. It is a simple framework for a problem- solving used in the aircraft industry. You can read about it at the link provided below. 

https://hfacs.com/hfacs-framework.html
 Following is a detailed PDF to help you answer the second question. 

https://dvikan.no/ntnu-studentserver/reports/A%20Human%20Error%20Approach%20to%20Aviation%20Accident%20Analysis.pdf
This book details the application of the HFACS system to analyse air crashes. Refer to tables 3.1 (Pg. 52), 3.2 (Pg. 58), 3.3 (Pg. 64) and 3.4 (Pg. 69), along with the detailed trees provided on pages 71 and 93.
 

For understanding the detailed definitions of all the levels in the HFACs framework, one can also refer to the link given below

https://www.skybrary.aero/index.php/Human_Factors_Analysis_and_Classification_System_(HFACS)