Add files via upload
Case Study 2 (Big Data Consultant in Complex Data)
Case Study 2: Problem Statement
You are a Big Data Tools and Technologies sales consultant, working for a company, Complex Data. Complex Data recently released the latest version of its Big Data tool, Data Fire, which helps companies analyse data and generate insights rapidly.

 

At a conference, you meet the Head of the Data Analytics team of a famous retail chain, Home Universe, during the networking session. You strike up a conversation and realize that the Home Universe is using tools and technologies that are not serving their needs; specifically, they are using Data Churn to analyse data and generate insights. The Data Analytics team has to respond to requests for insights, reports, and dashboards from different departments in the Home Universe. Given that these requests are for addressing urgent business needs, the team is usually under time pressure and faces a heavy workload.

 

You know that Data Churn is not only outdated but is also very slow and difficult to use compared with Data Fire, which is a state-of-the-art, easy-to-use solution with rapid response times; however, from a cost perspective, Data Churn costs significantly less than Data Fire.

 

Instead of jumping directly into pitching Data Fire to the Head of Data Analytics for Home Universe, your job now is to build the implications of using Data Churn so that the overall costs are reflected as much higher than in the case of Data Fire.

 

Based on this information, your task is to complete the blank spaces "-----" with suitable questions that the complex data team could ask the team of Home Universe during that part in the conversation in the attached word file given below. You can refer to the answers mentioned by the Home Universe team to get an understanding of what questions the Complex Data team could ask at that point in the conversation. Fill in the banks with the questions you feel are the most accurate to the situation.

 

Final Output Expected
Download the word file given below. You can fill in the questions in your copy of word file and submit back the complete conversation as a PDF file.