Add files via upload
Case Study 3: Problem Statement
Starbucks is the world’s largest coffeehouse chain, operating in over 30,000 locations and more than 70 countries. Use the reference material provided below and answer the subsequent questions on the marketing mix of Starbucks in India using the 7Ps approach.

 

The reference material is to help you answer the questions related to the 7Ps, which are as follows: 

Product
Price
Promotion
Place
People
Physical evidence
Process
We have designed a set of questions that will help you understand the business and the operations related to Starbucks. You need to refer to the reference material provided and answer the questions by researching the answers based on the reference material provided.

 

Reading Material and Final Output Expected
The following document consists of the references you would need for answering the questions related to each of the 7 Ps asked just below the references. You can either read all the information and then answer the questions or take a look at the questions and do your research to find the information.