Add files via upload
Case Study 4: Problem Statement
WalStore, a supermarket chain, has successfully grown from a few stores in just one city in the U.S. to over a hundred stores spread across 10 cities. As they are gearing up for the next wave of expansion to become a nationwide supermarket chain, they want to optimise their operations in order to ensure cost efficiencies. They have engaged a leading consulting firm, McBride & Co., to make recommendations about how WalStore’s end-to-end operations can be optimised.

 

In Phase 1, the McBride consultants plan to spend a couple of weeks gathering facts and data about WalStore’s operations. McBride & Co. already have a guide on supermarket operations from earlier assignments with other clients. Following is an extract from the guide, describing the various operational steps for any supermarket.

 

Your task for this question is to go through the extract below and then use the 5W+H framework to come up with a list of questions that will help McBride & Co. gather relevant facts and data about WalsStore’s current state of operations.

 

“McBride’s Guide To Supermarket Operations”
Six major entities are involved in the end-to-end operations for any supermarket. These entities are described as follows:

Suppliers: They produce goods that are to be sold in the supermarket. It should be noted that supermarkets do not produce goods themselves. They are in the service of retailing goods produced by suppliers. Depending upon the type of supermarket and their customer needs, a supermarket can have either a concentrated set of suppliers or a much diverse set of suppliers.
Warehouse transporters: They move the goods produced by suppliers to the concerned supermarket’s warehouses.
Warehouses: This is where goods that are produced by suppliers are stocked for further distribution. The locations of these warehouses, their size, operations (robotised or manual), etc. are thoroughly investigated.
Warehouse-to-store transporters: These transporters move the goods stocked in the warehouses to the stores where they are sold to customers.
Stores: Stores receive goods from transporters and stock them in their buffer supply area, from where they replenish the stock on the supermarket shelves.
Customers: Customers visit the stores, examine the goods stocked on the shelves and make the final purchase as per their needs.
The six entities mentioned above undertake steps, as depicted pictorially below, to make the goods produced by suppliers available to customers in stores.

Flow of the Operations between the 6 entities
Flow of the Operations between the 6 entities
Understand the scenario given above and in the document below a table is given. It consists of some sample questions and some spaces with “-------” in each question for the 6 entities given below. You have to fill these spaces with suitable questions that you think is a valid question for each of What, How, Who, When, Where? interrogation for each of the 6 operations of the WalStore i.e.

Supplier
Warehouse transporter
Warehouse
Warehouse-to-store transporter
Store
Customer